修改软件默认安装盘符.bat: Support x64 OS, fix default drive
x64 OS has a dedicated path for 32-bit programs and can not be altered
by the previous version of this batch program.
The default parameter can work normally now. To avoid potential
unintended modifications, the default drive is set to the system drive.
The program now tells user to click "Yes" or enter the password of
administrator in UAC dialog box.